requirements.txt
ijson
simplejson
pandas
nltk
numpy
scipy
matplotlib
scikit-learn
statsmodels
textblob
google-cloud-storage
google-cloud-secret-manager